Some minor code cleanup

+++ app/src/main/java/org/hexpresso/elm327/commands/protocol/ReadInputVoltageCommand.java
@@ -19,6 +19,6 @@ public class ReadInputVoltageCommand extends AbstractCommand {
         if (voltage == null) {
             voltage = new Double(getResponse().getLines().get(0));
         }
-        return voltage.doubleValue();
+        return voltage;
     }
 }
